+++ R/Population.R
@@ -288,6 +288,7 @@ Population <- R6Class(
         ), by = c(Ind$get_hid_col())] %>%
         # identify relationships
         .[, `:=`(
+          n_members = sapply(members, length),
           couple_hh = purrr::map2_lgl(members, partners, ~ {any(.y %in% .x)}),
           with_children = purrr::map2_lgl(members, parents, ~ {any(.y %in% .x)})
         )] %>%
@@ -306,7 +307,12 @@ Population <- R6Class(
           by.y = Ind$get_hid_col(),
           sort = FALSE,
           allow.cartesian = FALSE
-        )
+        ) %>%
+        # if there are individuals that don't belong to any household they would all
+        # be added into id:NA, so i thin
+        .[!is.na(id), ]
+
+
 
       checkmate::assert_character(household_type[["household_type"]], any.missing = FALSE)